requirements.txt
ultralytics==8.3.74
opencv-python==4.8
gradio==5.20.1
cvzone==1.6.1
pillow==11.1
numpy==1.24.3